backend/requirements.txt
django
django-rest-framework
Celery
django-cors-headers